rename config to cfg
