+++ extensions/pyRevitTools.extension/pyRevit.tab/Drawing Set.panel/views.stack/Legends.pulldown/FilterLegend.pushbutton/legend_cfg.py
